+++ src/object/blinnphong.cpp
@@ -28,7 +28,7 @@ Vector3f BlinnPhong::brdf(const Vector3f &point, const Vector3f &dir, const Vect
 	Vector3f thetaN = dir.normalize();
 	Vector3f Hn = (Ln + thetaN).normalize();
 
-	double i = powf((Nn*Hn), roughness);
+	float i = powf((float) (Nn*Hn), (float) roughness);
 
 	Vector3f ret = getSpecular()*i*specColor*specIntensity + getDiffuse()*getColor();
